telegram_bot/requirements.txt
python-telegram-bot[job-queue]==20.7
sqlalchemy==2.0.23
requests==2.31.0